fix: resolve native binding issue for lightningcss
Configure next.config.mjs, .npmrc, and postcss.config.js to use JS fallback for lightningcss.
