+++ .npmrc
@@ -9,8 +9,11 @@ loglevel=error
 shrinkwrap=false
 optional=false
 save-dev=false
-
 cache=/tmp/npm-cache
 prefer-dedupe=true
 legacy-bundling=false
 global-style=false
+# Disable native optional dependencies that don't have prebuilt binaries
+ignore-optional=true
+# Force lightningcss to use JS fallback
+LIGHTNINGCSS_WASM=true
