--- next.config.js
@@ -1,80 +1,0 @@
-const path = require("path")
-
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  turbopack: {
-    root: path.resolve(__dirname, ".."),
-  },
-  reactStrictMode: true,
-  poweredByHeader: false,
-
-  eslint: {
-    ignoreDuringBuilds: true,
-  },
-
-  typescript: {
-    ignoreBuildErrors: true,
-  },
-
-  images: {
-    remotePatterns: [
-      {
-        protocol: "https",
-        hostname: "**",
-      },
-      {
-        protocol: "http",
-        hostname: "**",
-      },
-    ],
-    unoptimized: true,
-  },
-
-  env: {
-    NEXT_PUBLIC_DEMO_MODE: process.env.NEXT_PUBLIC_DEMO_MODE || "true",
-    NEXT_PUBLIC_STATIC_MODE: process.env.NEXT_PUBLIC_STATIC_MODE || "false",
-    NEXT_PUBLIC_APP_VERSION: "1.0.0",
-  },
-
-  trailingSlash: false,
-  productionBrowserSourceMaps: false,
-
-  experimental: {
-    serverActions: {
-      bodySizeLimit: "10mb",
-    },
-  },
-
-  async headers() {
-    return [
-      {
-        source: "/(.*)",
-        headers: [
-          { key: "X-DNS-Prefetch-Control", value: "on" },
-          { key: "X-Frame-Options", value: "SAMEORIGIN" },
-          { key: "X-XSS-Protection", value: "1; mode=block" },
-          { key: "X-Content-Type-Options", value: "nosniff" },
-          { key: "Referrer-Policy", value: "strict-origin-when-cross-origin" },
-        ],
-      },
-      {
-        source: "/api/:path*",
-        headers: [
-          { key: "Cache-Control", value: "no-store, max-age=0" },
-        ],
-      },
-    ]
-  },
-
-  async redirects() {
-    return [
-      {
-        source: "/home",
-        destination: "/",
-        permanent: true,
-      },
-    ]
-  },
-}
-
-module.exports = nextConfig
